Move all non-dask/xarray methods and functions to base
