+++ pyresample/bilinear/__init__.py
@@ -32,9 +32,11 @@ http://www.ahinson.com/algorithms_general/Sections/InterpolationRegression/Inter
 
 import numpy as np
 from pyresample._spatial_mp import Proj
+from pyresample import data_reduce, geometry
 import warnings
 
 from pyresample import kd_tree
+from pykdtree.kdtree import KDTree
 
 
 def resample_bilinear(data, source_geo_def, target_area_def, radius=50e3,
@@ -240,10 +242,10 @@ def get_bil_info(source_geo_def, target_area_def, radius=50e3, neighbours=32,
     proj = Proj(target_area_def.proj_str)
 
     # Get output x/y coordinates
-    out_x, out_y = _get_output_xy(target_area_def, proj)
+    out_x, out_y = _get_output_xy_masked(target_area_def, proj)
 
     # Get input x/y coordinates
-    in_x, in_y = _get_input_xy(source_geo_def, proj, input_idxs, idx_ref)
+    in_x, in_y = _get_input_xy_masked(source_geo_def, proj, input_idxs, idx_ref)
 
     # Get the four closest corner points around each output location
     pt_1, pt_2, pt_3, pt_4, idx_ref = \
@@ -259,141 +261,6 @@ def get_bil_info(source_geo_def, target_area_def, radius=50e3, neighbours=32,
         s__ = np.ma.masked_where(mask, s__)
 
     return t__, s__, input_idxs, idx_ref
-
-
-def _get_fractional_distances(pt_1, pt_2, pt_3, pt_4, out_x, out_y):
-    """Calculate vertical and horizontal fractional distances t and s."""
-    # General case, ie. where the the corners form an irregular rectangle
-    t__, s__ = _get_fractional_distances_irregular(pt_1, pt_2, pt_3, pt_4, out_y, out_x)
-
-    # Cases where verticals are parallel
-    idxs = np.isnan(t__) | np.isnan(s__)
-    # Remove extra dimensions
-    idxs = idxs.ravel()
-
-    if np.any(idxs):
-        t__[idxs], s__[idxs] = _get_fractional_distances_uprights_parallel(
-                pt_1[idxs, :], pt_2[idxs, :],
-                pt_3[idxs, :], pt_4[idxs, :],
-                out_y[idxs], out_x[idxs])
-
-    # Cases where both verticals and horizontals are parallel
-    idxs = np.isnan(t__) | np.isnan(s__)
-    # Remove extra dimensions
-    idxs = idxs.ravel()
-    if np.any(idxs):
-        t__[idxs], s__[idxs] = _get_fractional_distances_parallellogram(
-            pt_1[idxs, :], pt_2[idxs, :], pt_3[idxs, :],
-            out_y[idxs], out_x[idxs])
-
-    with np.errstate(invalid='ignore'):
-        idxs = (t__ < 0) | (t__ > 1) | (s__ < 0) | (s__ > 1)
-    t__[idxs] = np.nan
-    s__[idxs] = np.nan
-
-    return t__, s__
-
-
-def _get_fractional_distances_irregular(pt_1, pt_2, pt_3, pt_4, out_y, out_x):
-    """Get parameters for the case where none of the sides are parallel."""
-    # Get parameters for the quadratic equation
-    a__, b__, c__ = _calc_abc(pt_1, pt_2, pt_3, pt_4, out_y, out_x)
-
-    # Get the valid roots from interval [0, 1]
-    t__ = _solve_quadratic(a__, b__, c__, min_val=0., max_val=1.)
-
-    # Calculate parameter s
-    s__ = _solve_another_fractional_distance(t__, pt_1[:, 1], pt_3[:, 1],
-                                             pt_2[:, 1], pt_4[:, 1], out_y)
-
-    return t__, s__
-
-
-def _get_fractional_distances_uprights_parallel(pt_1, pt_2, pt_3, pt_4, out_y, out_x):
-    """Get parameters for the case where uprights are parallel."""
-    # Get parameters for the quadratic equation
-    a__, b__, c__ = _calc_abc(pt_1, pt_3, pt_2, pt_4, out_y, out_x)
-
-    # Get the valid roots from interval [0, 1]
-    s__ = _solve_quadratic(a__, b__, c__, min_val=0., max_val=1.)
-
-    # Calculate parameter t
-    t__ = _solve_another_fractional_distance(s__, pt_1[:, 1], pt_2[:, 1],
-                                             pt_3[:, 1], pt_4[:, 1], out_y)
-
-    return t__, s__
-
-
-def _get_fractional_distances_parallellogram(pt_1, pt_2, pt_3, out_y, out_x):
-    """Get parameters for the case where uprights are parallel."""
-    # Pairwise longitudal separations between reference points
-    x_21 = pt_2[:, 0] - pt_1[:, 0]
-    x_31 = pt_3[:, 0] - pt_1[:, 0]
-
-    # Pairwise latitudal separations between reference points
-    y_21 = pt_2[:, 1] - pt_1[:, 1]
-    y_31 = pt_3[:, 1] - pt_1[:, 1]
-
-    t__ = (x_21 * (out_y - pt_1[:, 1]) - y_21 * (out_x - pt_1[:, 0])) / \
-          (x_21 * y_31 - y_21 * x_31)
-    with np.errstate(invalid='ignore'):
-        idxs = (t__ < 0.) | (t__ > 1.)
-    t__[idxs] = np.nan
-
-    s__ = (out_x - pt_1[:, 0] + x_31 * t__) / x_21
-
-    with np.errstate(invalid='ignore'):
-        idxs = (s__ < 0.) | (s__ > 1.)
-    s__[idxs] = np.nan
-
-    return t__, s__
-
-
-def _solve_another_fractional_distance(f__, y_1, y_2, y_3, y_4, out_y):
-    """Solve parameter t__ from s__, or vice versa.
-
-    For solving s__, switch order of y_2 and y_3.
-
-    """
-    y_21 = y_2 - y_1
-    y_43 = y_4 - y_3
-
-    with np.errstate(divide='ignore'):
-        g__ = ((out_y - y_1 - y_21 * f__) /
-               (y_3 + y_43 * f__ - y_1 - y_21 * f__))
-
-    # Limit values to interval [0, 1]
-    with np.errstate(invalid='ignore'):
-        idxs = (g__ < 0) | (g__ > 1)
-    g__[idxs] = np.nan
-
-    return g__
-
-
-def _calc_abc(pt_1, pt_2, pt_3, pt_4, out_y, out_x):
-    """Calculate coefficients for quadratic equation for irregular and upright cases.
-
-    For _get_fractional_distances_uprights switch order of pt_2 and pt_3.
-
-    """
-    # Pairwise longitudal separations between reference points
-    x_21 = pt_2[:, 0] - pt_1[:, 0]
-    x_31 = pt_3[:, 0] - pt_1[:, 0]
-    x_42 = pt_4[:, 0] - pt_2[:, 0]
-
-    # Pairwise latitudal separations between reference points
-    y_21 = pt_2[:, 1] - pt_1[:, 1]
-    y_31 = pt_3[:, 1] - pt_1[:, 1]
-    y_42 = pt_4[:, 1] - pt_2[:, 1]
-
-    a__ = x_31 * y_42 - y_31 * x_42
-    b__ = out_y * (x_42 - x_31) - out_x * (y_42 - y_31) + \
-        x_31 * pt_2[:, 1] - y_31 * pt_2[:, 0] + \
-        y_42 * pt_1[:, 0] - x_42 * pt_1[:, 1]
-    c__ = out_y * x_21 - out_x * y_21 + pt_1[:, 0] * pt_2[:, 1] - \
-        pt_2[:, 0] * pt_1[:, 1]
-
-    return a__, b__, c__
 
 
 def _mask_coordinates(lons, lats):
@@ -414,104 +281,7 @@ def _mask_coordinates(lons, lats):
     return lons, lats
 
 
-def _get_corner(stride, valid, in_x, in_y, idx_ref):
-    """Get closest set of coordinates from the *valid* locations."""
-    # Find the closest valid pixels, if any
-    idxs = np.argmax(valid, axis=1)
-    # Check which of these were actually valid
-    invalid = np.invert(np.max(valid, axis=1))
-
-    # Replace invalid points with np.nan
-    x__ = in_x[stride, idxs]
-    x__[invalid] = np.nan
-    y__ = in_y[stride, idxs]
-    y__[invalid] = np.nan
-    idx = idx_ref[stride, idxs]
-
-    return x__, y__, idx
-
-
-def _get_bounding_corners(in_x, in_y, out_x, out_y, neighbours, idx_ref):
-    """Get four closest locations from (in_x, in_y).
-
-    These locations will form a bounding rectangle around the requested location given
-    by (out_x, out_y).
-
-    """
-    # Find four closest pixels around the target location
-
-    # Tile output coordinates to same shape as neighbour info
-    out_x_tile = np.tile(out_x, (neighbours, 1)).T
-    out_y_tile = np.tile(out_y, (neighbours, 1)).T
-
-    # Get differences in both directions
-    x_diff = out_x_tile - in_x
-    y_diff = out_y_tile - in_y
-
-    stride = np.arange(x_diff.shape[0])
-
-    # Upper left source pixel
-    valid = (x_diff > 0) & (y_diff < 0)
-    x_1, y_1, idx_1 = _get_corner(stride, valid, in_x, in_y, idx_ref)
-
-    # Upper right source pixel
-    valid = (x_diff < 0) & (y_diff < 0)
-    x_2, y_2, idx_2 = _get_corner(stride, valid, in_x, in_y, idx_ref)
-
-    # Lower left source pixel
-    valid = (x_diff > 0) & (y_diff > 0)
-    x_3, y_3, idx_3 = _get_corner(stride, valid, in_x, in_y, idx_ref)
-
-    # Lower right source pixel
-    valid = (x_diff < 0) & (y_diff > 0)
-    x_4, y_4, idx_4 = _get_corner(stride, valid, in_x, in_y, idx_ref)
-
-    # Combine sorted indices to idx_ref
-    idx_ref = np.vstack((idx_1, idx_2, idx_3, idx_4)).T
-
-    return (np.vstack((x_1, y_1)).T, np.vstack((x_2, y_2)).T,
-            np.vstack((x_3, y_3)).T, np.vstack((x_4, y_4)).T, idx_ref)
-
-
-def _solve_quadratic(a__, b__, c__, min_val=0.0, max_val=1.0):
-    """Solve quadratic equation.
-
-    Return the valid roots from interval [*min_val*, *max_val*].
-
-    """
-
-    def int_and_float_to_numpy(val):
-        if not isinstance(val, np.ndarray):
-            if isinstance(val, (int, float)):
-                val = [val]
-            val = np.array(val)
-        return val
-
-    a__ = int_and_float_to_numpy(a__)
-    b__ = int_and_float_to_numpy(b__)
-    c__ = int_and_float_to_numpy(c__)
-
-    discriminant = b__ * b__ - 4 * a__ * c__
-
-    # Solve the quadratic polynomial
-    with np.errstate(invalid='ignore', divide='ignore'):
-        x_1 = (-b__ + np.sqrt(discriminant)) / (2 * a__)
-        x_2 = (-b__ - np.sqrt(discriminant)) / (2 * a__)
-
-    # Find valid solutions, ie. 0 <= t <= 1
-    x__ = x_1.copy()
-    with np.errstate(invalid='ignore'):
-        idxs = (x_1 < min_val) | (x_1 > max_val)
-    x__[idxs] = x_2[idxs]
-
-    with np.errstate(invalid='ignore'):
-        idxs = (x__ < min_val) | (x__ > max_val)
-    x__[idxs] = np.nan
-
-    return x__
-
-
-def _get_output_xy(target_area_def, proj):
+def _get_output_xy_masked(target_area_def, proj):
     """Get x/y coordinates of the target grid."""
     # Read output coordinates
     out_lons, out_lats = target_area_def.get_lonlats()
@@ -529,7 +299,7 @@ def _get_output_xy(target_area_def, proj):
     return out_x, out_y
 
 
-def _get_input_xy(source_geo_def, proj, input_idxs, idx_ref):
+def _get_input_xy_masked(source_geo_def, proj, input_idxs, idx_ref):
     """Get x/y coordinates for the input area and reduce the data."""
     in_lons, in_lats = source_geo_def.get_lonlats()
 
@@ -660,7 +430,14 @@ class BilinearBase(object):
             self._create_empty_bil_info()
 
     def _create_empty_bil_info(self):
-        raise NotImplementedError
+        self._valid_input_index = np.ones(self._source_geo_def.size, dtype=np.bool)
+        self._index_array = np.ones((self._target_geo_def.size, 4), dtype=np.int32)
+        self.bilinear_s = np.nan * np.zeros(self._target_geo_def.size)
+        self.bilinear_t = np.nan * np.zeros(self._target_geo_def.size)
+        self.slices_x = np.zeros((self._target_geo_def.size, 4), dtype=np.int32)
+        self.slices_y = np.zeros((self._target_geo_def.size, 4), dtype=np.int32)
+        self.out_coords_x, self.out_coords_y = self._target_geo_def.get_proj_vectors()
+        self.mask_slices = self._index_array >= self._source_geo_def.size
 
     def _get_target_lonlats(self):
         self._target_lons, self._target_lats = self._target_geo_def.get_lonlats()
@@ -673,23 +450,77 @@ class BilinearBase(object):
         index_array, _ = self._query_resample_kdtree()
         self._index_array = self._reduce_index_array(index_array)
 
-    def _query_resample_kdtree(self):
-        raise NotImplementedError
+    def _query_resample_kdtree(self,
+                               reduce_data=True):
+        """Query kd-tree on slice of target coordinates."""
+        res = query_no_distance(self._target_lons, self._target_lats,
+                                self._valid_output_index, self._resample_kdtree,
+                                self._neighbours, self._epsilon,
+                                self._radius_of_influence)
+        return res, None
 
     def _reduce_index_array(self, index_array):
-        raise NotImplementedError
+        input_size = np.sum(self._valid_input_index)
+        index_mask = index_array == input_size
+        return np.where(index_mask, 0, index_array)
 
     def _get_fractional_distances(self):
-        raise NotImplementedError
+        out_x, out_y = self._get_output_xy()
+        # Get the four closest corner points around each output location
+        pt_1, pt_2, pt_3, pt_4, self._index_array = \
+            _get_bounding_corners(*self._get_input_xy(),
+                                  out_x, out_y,
+                                  self._neighbours, self._index_array)
+        self.bilinear_t, self.bilinear_s = _get_fractional_distances(
+            pt_1, pt_2, pt_3, pt_4, out_x, out_y)
+
+    def _get_output_xy(self):
+        return _get_output_xy(self._target_geo_def)
+
+    def _get_input_xy(self):
+        return _get_input_xy(self._source_geo_def,
+                             Proj(self._target_geo_def.proj_str),
+                             self._valid_input_index, self._index_array)
 
     def _get_target_proj_vectors(self):
-        raise NotImplementedError
+        try:
+            self.out_coords_x, self.out_coords_y = self._target_geo_def.get_proj_vectors()
+        except AttributeError:
+            pass
 
     def _get_slices(self):
-        raise NotImplementedError
+        shp = self._source_geo_def.shape
+        cols, lines = np.meshgrid(np.arange(shp[1]),
+                                  np.arange(shp[0]))
+
+        self.slices_y, self.slices_x = array_slice_for_multiple_arrays(
+            self._index_array,
+            array_slice_for_multiple_arrays(
+                self._valid_input_index,
+                (np.ravel(lines), np.ravel(cols))
+            )
+        )
+        self.mask_slices = self._index_array >= self._source_geo_def.size
 
     def _create_resample_kdtree(self):
-        raise NotImplementedError
+        """Set up kd tree on input."""
+        valid_input_index, input_coords = self._get_valid_input_index_and_input_coords()
+        kdtree = None
+        if input_coords.size:
+            kdtree = KDTree(input_coords)
+        return valid_input_index, kdtree
+
+    def _get_valid_input_index_and_input_coords(self):
+        valid_input_index, source_lons, source_lats = \
+            get_valid_input_index(self._source_geo_def,
+                                  self._target_geo_def,
+                                  self._reduce_data,
+                                  self._radius_of_influence)
+        input_coords = lonlat2xyz(source_lons, source_lats)
+        valid_input_index = np.ravel(valid_input_index)
+        input_coords = input_coords[valid_input_index, :].astype(np.float)
+
+        return valid_input_index, input_coords
 
     def get_sample_from_bil_info(self, data, fill_value=None, output_shape=None):
         """Resample using pre-computed resampling LUTs."""
@@ -709,7 +540,7 @@ class BilinearBase(object):
                 p_4 * s__ * t__)
 
     def _slice_data(self, data, fill_value):
-        raise NotImplementedError
+        return get_slicer(data)(data, self.slices_x, self.slices_y, self.mask_slices, fill_value)
 
     def _finalize_output_data(self, data, res, fill_value):
         raise NotImplementedError
@@ -729,3 +560,377 @@ def _check_fill_value(fill_value, dtype):
             fill_value = int(fill_value)
 
     return fill_value
+
+
+def _get_output_xy(target_geo_def):
+    out_x, out_y = target_geo_def.get_proj_coords()
+    return np.ravel(out_x),  np.ravel(out_y)
+
+
+def _get_input_xy(source_geo_def, proj, valid_input_index, index_array):
+    """Get x/y coordinates for the input area and reduce the data."""
+    return proj(
+        *array_slice_for_multiple_arrays(
+            index_array,
+            array_slice_for_multiple_arrays(
+                valid_input_index,
+                mask_coordinates(*_get_raveled_lonlats(source_geo_def))
+            )
+        )
+    )
+
+
+def array_slice_for_multiple_arrays(idxs, data):
+    """Slices multiple arrays using the same indices."""
+    return [d[idxs] for d in data]
+
+
+def mask_coordinates(lons, lats):
+    """Mask invalid coordinate values."""
+    idxs = (find_indices_outside_min_and_max(lons, -180., 180.) |
+            find_indices_outside_min_and_max(lats, -90., 90.))
+    return _np_where_for_multiple_arrays(idxs, (np.nan, np.nan), (lons, lats))
+
+
+def find_indices_outside_min_and_max(data, min_val, max_val):
+    """Return array indices outside the given minimum and maximum values."""
+    return (data < min_val) | (data > max_val)
+
+
+def _get_raveled_lonlats(geo_def):
+    lons, lats = geo_def.get_lonlats()
+    if lons.size == 0 or lats.size == 0:
+        raise ValueError('Cannot resample empty data set')
+    elif lons.size != lats.size or lons.shape != lats.shape:
+        raise ValueError('Mismatch between lons and lats')
+
+    return np.ravel(lons), np.ravel(lats)
+
+
+def _get_bounding_corners(in_x, in_y, out_x, out_y, neighbours, index_array):
+    """Get bounding corners.
+
+    Get four closest locations from (in_x, in_y) so that they form a
+    bounding rectangle around the requested location given by (out_x,
+    out_y).
+
+    """
+    # Find four closest pixels around the target location
+
+    stride, valid_corners = _get_stride_and_valid_corner_indices(
+        out_x, out_y, in_x, in_y, neighbours)
+    res = []
+    indices = []
+    for valid in valid_corners:
+        x__, y__, idx = _get_corner(stride, valid, in_x, in_y, index_array)
+        res.append(np.transpose(np.vstack((x__, y__))))
+        indices.append(idx)
+    res.append(np.transpose(np.vstack(indices)))
+
+    return res
+
+
+def _get_fractional_distances(pt_1, pt_2, pt_3, pt_4, out_x, out_y):
+    """Calculate vertical and horizontal fractional distances t and s."""
+    # General case, ie. where the the corners form an irregular rectangle
+    t__, s__ = _invalid_s_and_t_to_nan(
+        *_get_fractional_distances_irregular(pt_1, pt_2, pt_3, pt_4, out_y, out_x)
+    )
+    # Update where verticals are parallel
+    t__, s__ = _update_fractional_distances(
+        _get_fractional_distances_uprights_parallel,
+        t__, s__, (pt_1, pt_2, pt_3, pt_4), out_x, out_y
+    )
+    # Update where both verticals and horizontals are parallel
+    t__, s__ = _update_fractional_distances(
+        _get_fractional_distances_parallellogram,
+        t__, s__, (pt_1, pt_2, pt_3), out_x, out_y
+    )
+
+    return t__, s__
+
+
+def _invalid_s_and_t_to_nan(t__, s__):
+    return _np_where_for_multiple_arrays(
+        (find_indices_outside_min_and_max(t__, 0, 1) |
+         find_indices_outside_min_and_max(s__, 0, 1)),
+        (np.nan, np.nan),
+        (t__, s__))
+
+
+def _get_fractional_distances_irregular(pt_1, pt_2, pt_3, pt_4, out_y, out_x):
+    """Get parameters for the case where none of the sides are parallel."""
+    # Get the valid roots from interval [0, 1]
+    t__ = _solve_quadratic(
+        *_calc_abc(pt_1, pt_2, pt_3, pt_4, out_y, out_x),
+        min_val=0., max_val=1.)
+
+    # Calculate parameter s
+    s__ = _solve_another_fractional_distance(t__, pt_1[:, 1], pt_3[:, 1],
+                                             pt_2[:, 1], pt_4[:, 1], out_y)
+
+    return t__, s__
+
+
+def _solve_quadratic(a__, b__, c__, min_val=0.0, max_val=1.0):
+    """Solve quadratic equation.
+
+    Solve quadratic equation and return the valid roots from interval
+    [*min_val*, *max_val*].
+
+    """
+    discriminant = b__ * b__ - 4 * a__ * c__
+
+    # Solve the quadratic polynomial
+    x_1 = (-b__ + np.sqrt(discriminant)) / (2 * a__)
+    x_2 = (-b__ - np.sqrt(discriminant)) / (2 * a__)
+
+    # Find valid solutions, ie. 0 <= t <= 1
+    x__ = np.where(
+        find_indices_outside_min_and_max(x_1, min_val, max_val),
+        x_2, x_1)
+
+    x__ = np.where(
+        find_indices_outside_min_and_max(x__, min_val, max_val),
+        np.nan, x__)
+
+    return x__
+
+
+def _calc_abc(pt_1, pt_2, pt_3, pt_4, out_y, out_x):
+    """Calculate coefficients for quadratic equation.
+
+    In this order of arguments used for _get_fractional_distances_irregular() and
+    _get_fractional_distances_uprights().  For _get_fractional_distances_uprights switch order of pt_2 and
+    pt_3.
+
+    """
+    # Pairwise longitudal separations between reference points
+    x_21 = pt_2[:, 0] - pt_1[:, 0]
+    x_31 = pt_3[:, 0] - pt_1[:, 0]
+    x_42 = pt_4[:, 0] - pt_2[:, 0]
+
+    # Pairwise latitudal separations between reference points
+    y_21 = pt_2[:, 1] - pt_1[:, 1]
+    y_31 = pt_3[:, 1] - pt_1[:, 1]
+    y_42 = pt_4[:, 1] - pt_2[:, 1]
+
+    a__ = x_31 * y_42 - y_31 * x_42
+    b__ = out_y * (x_42 - x_31) - out_x * (y_42 - y_31) + \
+        x_31 * pt_2[:, 1] - y_31 * pt_2[:, 0] + \
+        y_42 * pt_1[:, 0] - x_42 * pt_1[:, 1]
+    c__ = out_y * x_21 - out_x * y_21 + pt_1[:, 0] * pt_2[:, 1] - \
+        pt_2[:, 0] * pt_1[:, 1]
+
+    return a__, b__, c__
+
+
+def _solve_another_fractional_distance(f__, y_1, y_2, y_3, y_4, out_y):
+    """Solve parameter t__ from s__, or vice versa.
+
+    For solving s__, switch order of y_2 and y_3.
+    """
+    y_21 = y_2 - y_1
+    y_43 = y_4 - y_3
+
+    g__ = ((out_y - y_1 - y_21 * f__) /
+           (y_3 + y_43 * f__ - y_1 - y_21 * f__))
+
+    # Limit values to interval [0, 1]
+    g__ = np.where(
+        find_indices_outside_min_and_max(g__, 0, 1),
+        np.nan, g__)
+
+    return g__
+
+
+def _update_fractional_distances(func, t__, s__, points, out_x, out_y):
+    idxs = np.ravel(np.isnan(t__) | np.isnan(s__))
+    if np.any(idxs):
+        t__, s__ = _invalid_s_and_t_to_nan(
+            *_np_where_for_multiple_arrays(
+                idxs,
+                func(
+                    *points, out_y, out_x),
+                (t__, s__)
+            )
+        )
+    return t__, s__
+
+
+def _get_fractional_distances_uprights_parallel(pt_1, pt_2, pt_3, pt_4, out_y, out_x):
+    """Get parameters for the case where uprights are parallel."""
+    # Get the valid roots from interval [0, 1]
+    s__ = _solve_quadratic(
+        *_calc_abc(pt_1, pt_3, pt_2, pt_4, out_y, out_x),
+        min_val=0., max_val=1.)
+
+    # Calculate parameter t
+    t__ = _solve_another_fractional_distance(s__, pt_1[:, 1], pt_2[:, 1],
+                                             pt_3[:, 1], pt_4[:, 1], out_y)
+
+    return t__, s__
+
+
+def _get_fractional_distances_parallellogram(pt_1, pt_2, pt_3, out_y, out_x):
+    """Get parameters for the case where uprights are parallel."""
+    # Pairwise longitudal separations between reference points
+    x_21 = pt_2[:, 0] - pt_1[:, 0]
+    x_31 = pt_3[:, 0] - pt_1[:, 0]
+
+    # Pairwise latitudal separations between reference points
+    y_21 = pt_2[:, 1] - pt_1[:, 1]
+    y_31 = pt_3[:, 1] - pt_1[:, 1]
+
+    t__ = (x_21 * (out_y - pt_1[:, 1]) - y_21 * (out_x - pt_1[:, 0])) / \
+          (x_21 * y_31 - y_21 * x_31)
+    t__ = np.where(
+        find_indices_outside_min_and_max(t__, 0., 1.),
+        np.nan, t__)
+
+    s__ = (out_x - pt_1[:, 0] + x_31 * t__) / x_21
+    s__ = np.where(
+        find_indices_outside_min_and_max(s__, 0., 1.),
+        np.nan, s__)
+
+    return t__, s__
+
+
+def _get_stride_and_valid_corner_indices(out_x, out_y, in_x, in_y, neighbours):
+    out_x_tile = _tile_output_coordinate_vector(out_x, neighbours)
+    out_y_tile = _tile_output_coordinate_vector(out_y, neighbours)
+
+    # Get differences in both directions
+    x_diff = out_x_tile - in_x
+    y_diff = out_y_tile - in_y
+
+    return (np.arange(x_diff.shape[0]), (
+        (x_diff > 0) & (y_diff < 0),  # Upper left corners
+        (x_diff < 0) & (y_diff < 0),  # Upper right corners
+        (x_diff > 0) & (y_diff > 0),  # Lower left corners
+        (x_diff < 0) & (y_diff > 0))  # Lower right corners
+    )
+
+
+def _tile_output_coordinate_vector(vector, neighbours):
+    return np.reshape(np.tile(vector, neighbours), (neighbours, vector.size)).T
+
+
+def _get_corner(stride, valid, in_x, in_y, index_array):
+    """Get closest set of coordinates from the *valid* locations."""
+    x__, y__, idx = _slice_2d_with_stride_and_indices_for_multiple_arrays(
+        (in_x, in_y, index_array),
+        stride,
+        np.argmax(valid, axis=1)  # The closest valid locations
+    )
+    # Replace invalid points with np.nan
+    x__, y__ = _np_where_for_multiple_arrays(
+        np.invert(np.max(valid, axis=1)),
+        (np.nan, np.nan),
+        (x__, y__)
+    )
+
+    return x__, y__, idx
+
+
+def _slice_2d_with_stride_and_indices_for_multiple_arrays(arrays, stride, idxs):
+    return [arr[stride, idxs] for arr in arrays]
+
+
+def _np_where_for_multiple_arrays(idxs, values_for_idxs, otherwise_arrays):
+    return [np.where(idxs, values_for_idxs[i], arr) for i, arr in enumerate(otherwise_arrays)]
+
+
+def get_valid_input_index(source_geo_def,
+                          target_geo_def,
+                          reduce_data,
+                          radius_of_influence):
+    """Find indices of reduced input data."""
+    source_lons, source_lats = _get_raveled_lonlats(source_geo_def)
+
+    valid_input_index = np.invert(
+        find_indices_outside_min_and_max(source_lons, -180., 180.)
+        | find_indices_outside_min_and_max(source_lats, -90., 90.))
+
+    if reduce_data and _is_swath_to_grid_or_grid_to_grid(source_geo_def, target_geo_def):
+        valid_input_index &= _get_valid_indices_from_lonlat_boundaries(
+            target_geo_def, source_lons, source_lats, radius_of_influence)
+
+    return valid_input_index, source_lons, source_lats
+
+
+def _is_swath_to_grid_or_grid_to_grid(source_geo_def, target_geo_def):
+    return (isinstance(source_geo_def, geometry.CoordinateDefinition) and
+            isinstance(target_geo_def, (geometry.GridDefinition,
+                                        geometry.AreaDefinition))) or \
+           (isinstance(source_geo_def, (geometry.GridDefinition,
+                                        geometry.AreaDefinition)) and
+            isinstance(target_geo_def, (geometry.GridDefinition,
+                                        geometry.AreaDefinition)))
+
+
+def _get_valid_indices_from_lonlat_boundaries(
+        target_geo_def, source_lons, source_lats, radius_of_influence):
+    # Resampling from swath to grid or from grid to grid
+    lonlat_boundary = target_geo_def.get_boundary_lonlats()
+
+    # Combine reduced and legal values
+    return data_reduce.get_valid_index_from_lonlat_boundaries(
+        lonlat_boundary[0],
+        lonlat_boundary[1],
+        source_lons, source_lats,
+        radius_of_influence)
+
+
+def lonlat2xyz(lons, lats):
+    """Convert geographic coordinates to cartesian 3D coordinates."""
+    R = 6370997.0
+    lats = np.deg2rad(lats)
+    r_cos_lats = R * np.cos(lats)
+    lons = np.deg2rad(lons)
+    x_coords = r_cos_lats * np.cos(lons)
+    y_coords = r_cos_lats * np.sin(lons)
+    z_coords = R * np.sin(lats)
+
+    return np.stack(
+        (x_coords.ravel(), y_coords.ravel(), z_coords.ravel()), axis=-1)
+
+
+def get_slicer(data):
+    """Get slicing function for 2D or 3D arrays, depending on the data."""
+    if data.ndim == 2:
+        return _slice2d
+    elif data.ndim == 3:
+        return _slice3d
+    else:
+        raise ValueError
+
+
+def _slice2d(values, sl_x, sl_y, mask, fill_value):
+    # Slice 2D data
+    arr = values[(sl_y, sl_x)]
+    arr[(mask, )] = fill_value
+    return arr[:, 0], arr[:, 1], arr[:, 2], arr[:, 3]
+
+
+def _slice3d(values, sl_x, sl_y, mask, fill_value):
+    # Slice 3D data
+    arr = values[(slice(None), sl_y, sl_x)]
+    arr[(slice(None), mask)] = fill_value
+    return arr[:, :, 0], arr[:, :, 1], arr[:, :, 2], arr[:, :, 3]
+
+
+def query_no_distance(target_lons, target_lats,
+                      valid_output_index, kdtree, neighbours, epsilon, radius):
+    """Query the kdtree. No distances are returned."""
+    voir = np.ravel(valid_output_index)
+    target_lons_valid = np.ravel(target_lons)[voir]
+    target_lats_valid = np.ravel(target_lats)[voir]
+
+    _, index_array = kdtree.query(
+        lonlat2xyz(target_lons_valid, target_lats_valid),
+        k=neighbours,
+        eps=epsilon,
+        distance_upper_bound=radius)
+
+    return index_array
