+++ pyresample/test/test_bilinear.py
@@ -668,7 +668,7 @@ class TestXarrayBilinear(unittest.TestCase):
         resampler._get_slices()
         self.assertTrue(np.all(resampler.mask_slices))
 
-    @mock.patch('pyresample.bilinear.xarr.KDTree')
+    @mock.patch('pyresample.bilinear.KDTree')
     def test_create_resample_kdtree(self, KDTree):
         """Test that KDTree creation is called."""
         from pyresample.bilinear.xarr import XArrayResamplerBilinear
@@ -681,7 +681,7 @@ class TestXarrayBilinear(unittest.TestCase):
         self.assertEqual(vii.size, self.source_def.size)
         KDTree.assert_called_once()
 
-    @mock.patch('pyresample.bilinear.xarr.query_no_distance')
+    @mock.patch('pyresample.bilinear.query_no_distance')
     def test_query_resample_kdtree(self, qnd):
         """Test that query_no_distance is called in _query_resample_kdtree()."""
         from pyresample.bilinear.xarr import XArrayResamplerBilinear
@@ -715,9 +715,9 @@ class TestXarrayBilinear(unittest.TestCase):
     def test_mask_coordinates(self):
         """Test masking of invalid coordinates."""
         import dask.array as da
-        from pyresample.bilinear.xarr import _mask_coordinates
+        from pyresample.bilinear import mask_coordinates
 
-        lons, lats = _mask_coordinates(
+        lons, lats = mask_coordinates(
             da.from_array([-200., 0., 0., 0., 200.]),
             da.from_array([0., -100., 0, 100., 0.]))
         lons, lats = da.compute(lons, lats)
@@ -728,8 +728,8 @@ class TestXarrayBilinear(unittest.TestCase):
     def test_get_bounding_corners(self):
         """Test finding surrounding bounding corners."""
         import dask.array as da
-        from pyresample.bilinear.xarr import (_get_input_xy,
-                                              _get_bounding_corners)
+        from pyresample.bilinear.xarr import _get_input_xy
+        from pyresample.bilinear import _get_bounding_corners
         from pyresample._spatial_mp import Proj
         from pyresample import CHUNK_SIZE
 
@@ -758,8 +758,7 @@ class TestXarrayBilinear(unittest.TestCase):
     def test_get_corner(self):
         """Test finding the closest corners."""
         import dask.array as da
-        from pyresample.bilinear.xarr import (_get_corner,
-                                              _get_input_xy)
+        from pyresample.bilinear import _get_corner, _get_input_xy
         from pyresample import CHUNK_SIZE
         from pyresample._spatial_mp import Proj
 
@@ -791,12 +790,12 @@ class TestXarrayBilinear(unittest.TestCase):
         # bottom row of the area
         self.assertEqual(np.sum(np.isnan(x_3.compute())), 4)
 
-    @mock.patch('pyresample.bilinear.xarr._get_fractional_distances_parallellogram')
-    @mock.patch('pyresample.bilinear.xarr._get_fractional_distances_uprights_parallel')
-    @mock.patch('pyresample.bilinear.xarr._get_fractional_distances_irregular')
+    @mock.patch('pyresample.bilinear._get_fractional_distances_parallellogram')
+    @mock.patch('pyresample.bilinear._get_fractional_distances_uprights_parallel')
+    @mock.patch('pyresample.bilinear._get_fractional_distances_irregular')
     def test_get_fractional_distances(self, irregular, uprights, parallellogram):
         """Test that the three separate functions are called."""
-        from pyresample.bilinear.xarr import _get_fractional_distances
+        from pyresample.bilinear import _get_fractional_distances
         import dask.array as da
 
         # All valid values
@@ -870,7 +869,7 @@ class TestXarrayBilinear(unittest.TestCase):
 
     def test_get_fractional_distances_irregular(self):
         """Test calculations for irregular corner locations."""
-        from pyresample.bilinear.xarr import _get_fractional_distances_irregular
+        from pyresample.bilinear import _get_fractional_distances_irregular
 
         res = _get_fractional_distances_irregular(
             self.pts_irregular[0],
@@ -916,7 +915,7 @@ class TestXarrayBilinear(unittest.TestCase):
 
     def test_calc_abc(self):
         """Test calculation of quadratic coefficients."""
-        from pyresample.bilinear.xarr import _calc_abc
+        from pyresample.bilinear import _calc_abc
 
         # No np.nan inputs
         pt_1, pt_2, pt_3, pt_4 = self.pts_irregular
@@ -934,8 +933,8 @@ class TestXarrayBilinear(unittest.TestCase):
     def test_solve_quadratic(self):
         """Test solving quadratic equation."""
         import dask.array as da
-        from pyresample.bilinear.xarr import (_solve_quadratic,
-                                              _calc_abc)
+        from pyresample.bilinear import (_solve_quadratic,
+                                         _calc_abc)
 
         res = _solve_quadratic(1, 0, 0)
         self.assertEqual(res, 0.0)
@@ -957,7 +956,7 @@ class TestXarrayBilinear(unittest.TestCase):
 
     def test_query_no_distance(self):
         """Test KDTree querying."""
-        from pyresample.bilinear.xarr import query_no_distance
+        from pyresample.bilinear import query_no_distance
 
         kdtree = mock.MagicMock()
         kdtree.query.return_value = (1, 2)
@@ -971,23 +970,23 @@ class TestXarrayBilinear(unittest.TestCase):
 
     def test_get_valid_input_index(self):
         """Test finding valid indices for reduced input data."""
-        from pyresample.bilinear.xarr import _get_valid_input_index
+        from pyresample.bilinear import get_valid_input_index
 
         # Do not reduce data
-        vii, lons, lats = _get_valid_input_index(self.source_def,
-                                                 self.target_def,
-                                                 False, self.radius)
+        vii, lons, lats = get_valid_input_index(self.source_def,
+                                                self.target_def,
+                                                False, self.radius)
         self.assertEqual(vii.shape, (self.source_def.size, ))
         self.assertTrue(vii.dtype == np.bool)
         # No data has been reduced, whole input is used
-        self.assertTrue(vii.compute().all())
+        self.assertTrue(vii.all())
 
         # Reduce data
-        vii, lons, lats = _get_valid_input_index(self.source_def,
-                                                 self.target_def,
-                                                 True, self.radius)
+        vii, lons, lats = get_valid_input_index(self.source_def,
+                                                self.target_def,
+                                                True, self.radius)
         # 2700 valid input points
-        self.assertEqual(vii.compute().sum(), 2700)
+        self.assertEqual(vii.sum(), 2700)
 
     def test_create_empty_bil_info(self):
         """Test creation of empty bilinear info."""
@@ -1006,7 +1005,7 @@ class TestXarrayBilinear(unittest.TestCase):
 
     def test_lonlat2xyz(self):
         """Test conversion from geographic to cartesian 3D coordinates."""
-        from pyresample.bilinear.xarr import lonlat2xyz
+        from pyresample.bilinear import lonlat2xyz
         from pyresample import CHUNK_SIZE
 
         lons, lats = self.target_def.get_lonlats(chunks=CHUNK_SIZE)
@@ -1018,7 +1017,7 @@ class TestXarrayBilinear(unittest.TestCase):
 
 def test_check_fill_value():
     """Test that fill_value replacement/adjustment works."""
-    from pyresample.bilinear.xarr import _check_fill_value
+    from pyresample.bilinear import _check_fill_value
 
     # None + integer dtype -> 0
     assert _check_fill_value(None, np.uint8) == 0
